+++ sapp/pixelformats-sapp.c
@@ -471,7 +471,7 @@ static sg_range gen_pixels(sg_pixel_format fmt) {
         case SG_PIXELFORMAT_RGBA8:      gen_pixels_32(0xFFFFFFFF);  return (sg_range) {pixels, 8*8*4};
         case SG_PIXELFORMAT_RGBA8SN:    gen_pixels_32(0x7F7F7F7F); return (sg_range) {pixels, 8*8*4};
         case SG_PIXELFORMAT_BGRA8:      gen_pixels_32(0xFFFFFFFF); return (sg_range) {pixels, 8*8*4};
-        case SG_PIXELFORMAT_RGB10A2:    gen_pixels_32(0x3<<30 | 0x3FF<<20 | 0x3FF<<10 | 0x3FF); return (sg_range) {pixels, 8*8*4};
+        case SG_PIXELFORMAT_RGB10A2:    gen_pixels_32((uint32_t)(0x3<<30 | 0x3FF<<20 | 0x3FF<<10 | 0x3FF)); return (sg_range) {pixels, 8*8*4};
         case SG_PIXELFORMAT_RG11B10F:   gen_pixels_32(0x1E0<<22 | 0x3C0<<11 | 0x3C0); return (sg_range) {pixels, 8*8*4};
         case SG_PIXELFORMAT_RG32F:      gen_pixels_64(0x3F8000003F800000); return (sg_range) {pixels, 8*8*8};
         case SG_PIXELFORMAT_RGBA16:     gen_pixels_64(0xFFFFFFFFFFFFFFFF); return (sg_range) {pixels, 8*8*8};
